Restructure project with a new middleware folder
